Use ComponentRenderer for both page/layout components
